requirements.txt
openai
scipy
numpy
python-dotenv